requirements.txt
streamlit
pandas
google-play-scraper
langdetect
groq
google-generativeai
fpdf2
